+++ ChangeLog.md
@@ -1,4 +1,4 @@
-## 1.1.8
+## 1.2.0
 
 * Added Unit test
 * Added example for transient with external request
@@ -6,6 +6,7 @@
 * Improved code for shortcode and filter
 * Fixed Apigen and [git hook](https://github.com/Mte90/WordPress-Plugin-Boilerplate-Powered/wiki/Hook-for-generate-doc-and-pushing)
 * Moved code in class-admin-plugin-name.php in dedicated files
+* Moved activaction and deactivation on specific file in public/includes
 * Moved help-docs in admin/includes
 
 ## 1.1.7